Started creating a health script and pause/win/fail script
There is now a GameStateHandler script that is meant to facilitate easily switching between different game states, but is not completed yet. I need to add code for an actual pause button.

There is a Health script, with code for taking damage, healing and dying. There is a PlayerHealth derived class with different code for dying and respawning, and I need to make an EnemyHealth derived class for code such as pain/death animations.