The player can take damage now!
This was an error on my part. The code to check invincibility was referencing the wrong variable, and the player would constantly register as invincible. Now the player can take damage and die. Jump into the water and try it! However, I tried adding a 'Respawn' button to the main menu, but the player cannot click on it for some reason.